Added base for new collision systen
Updated dummy texture and dummy collision, updated slime collision, updated devourer of gods head and tail collision.
More mechanics based off this will be implemented later such as: Damaging enemies, giving each enemy a unique damage value, different parts dealing different damage etc.
