+++ Reign of Yharim/Assets/Materials/No Friction.physicsMaterial2D
@@ -8,4 +8,4 @@ PhysicsMaterial2D:
   m_PrefabAsset: {fileID: 0}
   m_Name: No Friction
   friction: 0
-  bounciness: 0.1
+  bounciness: 0.01
