--- Reign of Yharim/Assets/Prefabs/AstrumAureus.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 7f3ef1842f41c344aaa7fbd8e901c33d
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
