--- Reign of Yharim/Assets/Scripts/AI/AureusAI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8e6431c85f319244c9cf0ef964a6f304
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
