+++ Reign of Yharim/Assets/Scripts/CollisionSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: edc5666ea71c4eb4b91b56861da35f2f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
