+++ Reign of Yharim/Assets/Scripts/GroundDetection.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cab18c29b8fe9c04faa6cb8633d23f01
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
