+++ Reign of Yharim/Assets/Scripts/UtilityPrograms.cs
@@ -4,37 +4,8 @@ using UnityEngine;
 
 public class UtilityPrograms : MonoBehaviour
 {
-    public bool lockRotation = false;
-    public bool detectCollisionPlayer = false;
     void LateUpdate()
     {
-        if(lockRotation == true)
-            transform.rotation = Quaternion.Euler(0, 0, 0); //locks the rotation of attached objects
-        else
-            return;
-    }
-    private void OnTriggerStay2D(Collider2D collision)
-    {
-        if (detectCollisionPlayer == true)
-        {
-            if (collision.gameObject.layer == LayerMask.NameToLayer("Ground"))
-            {
-                PlayerAI.isGrounded = true;
-            }
-        }
-        else
-            return;
-    }
-    private void OnTriggerExit2D(Collider2D collision)
-    {
-        if (detectCollisionPlayer == true)
-        {
-            if (collision.gameObject.layer == LayerMask.NameToLayer("Ground"))
-            {
-                PlayerAI.isGrounded = false;
-            }
-        }
-        else
-            return;
+        transform.rotation = Quaternion.Euler(0, 0, 0); //locks the rotation of attached objects
     }
 }
