+++ Reign of Yharim/Assets/Textures/Player/Player.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8566ea2cd2b0ccf4a85d5d736e15114a
+guid: f03cb81e68b6d484a90a513fc1a49776
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
