+++ Reign of Yharim/ProjectSettings/TagManager.asset
@@ -5,6 +5,7 @@ TagManager:
   serializedVersion: 2
   tags:
   - AstrumAureus
+  - Hitbox
   layers:
   - Default
   - TransparentFX
@@ -16,7 +17,7 @@ TagManager:
   - Item
   - Ground
   - Background
-  - WalkThroughNPCs
+  - WalkThroughNPCPlayer
   - 
   - 
   -
