+++ Reign of Yharim/UserSettings/Layouts/default-2021.dwlt
@@ -19,7 +19,7 @@ MonoBehaviour:
     width: 1536
     height: 870
   m_ShowMode: 4
-  m_Title: Game
+  m_Title: Project
   m_RootView: {fileID: 2}
   m_MinSize: {x: 875, y: 300}
   m_MaxSize: {x: 10000, y: 10000}
@@ -119,7 +119,7 @@ MonoBehaviour:
   m_MinSize: {x: 300, y: 200}
   m_MaxSize: {x: 24288, y: 16192}
   vertical: 0
-  controlID: 69
+  controlID: 135
 --- !u!114 &6
 MonoBehaviour:
   m_ObjectHideFlags: 52
@@ -144,7 +144,7 @@ MonoBehaviour:
   m_MinSize: {x: 200, y: 200}
   m_MaxSize: {x: 16192, y: 16192}
   vertical: 1
-  controlID: 75
+  controlID: 123
 --- !u!114 &7
 MonoBehaviour:
   m_ObjectHideFlags: 52
@@ -169,7 +169,7 @@ MonoBehaviour:
   m_MinSize: {x: 200, y: 100}
   m_MaxSize: {x: 16192, y: 8096}
   vertical: 0
-  controlID: 76
+  controlID: 124
 --- !u!114 &8
 MonoBehaviour:
   m_ObjectHideFlags: 52
@@ -341,9 +341,9 @@ MonoBehaviour:
     m_VAllowExceedBaseRangeMax: 1
     m_ScaleWithWindow: 0
     m_HSlider: 0
-    m_VSlider: 0
+    m_VSlider: 1
     m_IgnoreScrollWheelUntilClicked: 0
-    m_EnableMouseInput: 0
+    m_EnableMouseInput: 1
     m_EnableSliderZoomHorizontal: 0
     m_EnableSliderZoomVertical: 0
     m_UniformScale: 1
@@ -354,18 +354,18 @@ MonoBehaviour:
       y: 21
       width: 920.4
       height: 491.2
-    m_Scale: {x: 1.1940521, y: 1.1940521}
-    m_Translation: {x: 460.2, y: 243.33453}
+    m_Scale: {x: 1.1790948, y: 1.1790948}
+    m_Translation: {x: 460.2, y: 246.07501}
     m_MarginLeft: 0
     m_MarginRight: 0
     m_MarginTop: 0
     m_MarginBottom: 0
     m_LastShownAreaInsideMargins:
       serializedVersion: 2
-      x: -385.41034
-      y: -203.78888
-      width: 770.8207
-      height: 411.37234
+      x: -390.2994
+      y: -208.69824
+      width: 780.5988
+      height: 416.59076
     m_MinimalGUI: 1
   m_defaultScale: 1
   m_LastWindowPixelSize: {x: 2301, y: 1280.5}
@@ -406,9 +406,9 @@ MonoBehaviour:
   m_SceneHierarchy:
     m_TreeViewState:
       scrollPos: {x: 0, y: 0}
-      m_SelectedIDs: 2238f8ff
-      m_LastClickedID: -509918
-      m_ExpandedIDs: 1ef9f7ff28f9f7ff2ef9f7ff0cfbffff806b00008c6b0000d86b0000e06b0000206c0000
+      m_SelectedIDs: 
+      m_LastClickedID: 0
+      m_ExpandedIDs: a44ff2ffa2b1f2ff88b3f2ff22b8f2ff7202f3ff8403f4ffaae5f4ffc0e8f4ffcae8f4ffd871f7ff267bf7ff547bf7ffa0cdf8ff14daf8ff1edaf8ff64dff8ff6edff8ff6e0cfdff3eefffff44efffff50efffff54effffff4faffff0afbffff0cfbffff086b00007a6b0000
       m_RenameOverlay:
         m_UserAcceptedRename: 0
         m_Name: 
@@ -713,9 +713,9 @@ MonoBehaviour:
   m_PlayAudio: 0
   m_AudioPlay: 0
   m_Position:
-    m_Target: {x: 11.758421, y: -6.067941, z: 806.7111}
+    m_Target: {x: 10.64405, y: -4.1828475, z: -0.14096718}
     speed: 2
-    m_Value: {x: 11.758421, y: -6.067941, z: 806.7111}
+    m_Value: {x: 10.64405, y: -4.1828475, z: -0.14096718}
   m_RenderMode: 0
   m_CameraMode:
     drawMode: 0
@@ -766,9 +766,9 @@ MonoBehaviour:
     speed: 2
     m_Value: {x: 0, y: 0, z: 0, w: 1}
   m_Size:
-    m_Target: 8.6971655
+    m_Target: 16.846388
     speed: 2
-    m_Value: 8.6971655
+    m_Value: 16.846388
   m_Ortho:
     m_Target: 1
     speed: 2
@@ -931,9 +931,9 @@ MonoBehaviour:
     m_IsLocked: 0
   m_FolderTreeState:
     scrollPos: {x: 0, y: 79}
-    m_SelectedIDs: e68c0000
-    m_LastClickedID: 36070
-    m_ExpandedIDs: 000000003870000034790000e68c000000ca9a3bffffff7f
+    m_SelectedIDs: 22700000
+    m_LastClickedID: 28706
+    m_ExpandedIDs: 000000000e700000227000002870000036700000067f0000948a0000a48b0000fc93000000ca9a3bffffff7f
     m_RenameOverlay:
       m_UserAcceptedRename: 0
       m_Name: 
@@ -961,7 +961,7 @@ MonoBehaviour:
     scrollPos: {x: 0, y: 0}
     m_SelectedIDs: 
     m_LastClickedID: 0
-    m_ExpandedIDs: 0000000038700000
+    m_ExpandedIDs: 000000000e700000
     m_RenameOverlay:
       m_UserAcceptedRename: 0
       m_Name: 
@@ -989,7 +989,7 @@ MonoBehaviour:
     m_SelectedInstanceIDs: 
     m_LastClickedInstanceID: 0
     m_HadKeyboardFocusLastEvent: 1
-    m_ExpandedInstanceIDs: c62300006c60000094610000d67d0000ca610000086d0000
+    m_ExpandedInstanceIDs: c62300006c60000094610000d67d0000ca610000086d0000dc6c0000
     m_RenameOverlay:
       m_UserAcceptedRename: 0
       m_Name: 
@@ -1077,7 +1077,7 @@ MonoBehaviour:
     m_OverlaysVisible: 1
   m_LockTracker:
     m_IsLocked: 0
-  m_LastSelectedObjectID: 27532
+  m_LastSelectedObjectID: -50604
 --- !u!114 &19
 MonoBehaviour:
   m_ObjectHideFlags: 52
